Adapt to latest libvxl

+++ src/chunk.h
@@ -23,6 +23,7 @@
 #include <stdint.h>
 #include <pthread.h>
 #include <stdbool.h>
+#include <stddef.h>
 
 #include "glx.h"
 #include "tesselator.h"
@@ -47,7 +48,7 @@ void chunk_block_update(int x, int y, int z);
 void chunk_update_all(void);
 void* chunk_generate(void* data);
 void chunk_generate_greedy(int start_x, int start_z, struct tesselator* tess, int* max_height);
-void chunk_generate_naive(struct libvxl_block* blocks, int count, uint32_t* solid, struct tesselator* tess,
+void chunk_generate_naive(struct libvxl_block* blocks, size_t count, size_t* solid, struct tesselator* tess,
 						  int* max_height, int ao);
 void chunk_rebuild_all(void);
 void chunk_draw_visible(void);
